refactor: validation codes has been updated in command handler classes

+++ src/Core/DapperORM.Application/Features/Commands/DeleteEvent/DeleteCategoryCommandHandler.cs
@@ -1,11 +1,11 @@
 ﻿using AutoMapper;
 using DapperORM.Application.Interfaces.Repositories;
-using DapperORM.Application.Validations.Common;
 using DapperORM.Application.Validations.Delete;
 using DapperORM.Domain.Common.Result;
 using DapperORM.Domain.Constants;
 using DapperORM.Domain.Entities;
 using MediatR;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -31,7 +31,9 @@ namespace DapperORM.Application.Features.Commands.DeleteEvent
         public Task<IResult> Handle(DeleteCategoryCommandRequest request, CancellationToken cancellationToken)
         {
             Category category = mapper.Map<Category>(request);
-            EntityValidator.Validate(validator, category);
+            var result = validator.Validate(category);
+            if (result.Errors.Any())
+                return Task.FromResult<IResult>(new ErrorResult(result.Errors.First().ErrorMessage));
             categoryRepository.Delete(category);
             return Task.FromResult<IResult>(new SuccessResult(ResultMessages.Category_Deleted));
         }
