+++ src/Core/DapperORM.Application/Features/Commands/DeleteEvent/DeleteProductCommandHandler.cs
@@ -1,11 +1,11 @@
 ﻿using AutoMapper;
 using DapperORM.Application.Interfaces.Repositories;
-using DapperORM.Application.Validations.Common;
 using DapperORM.Application.Validations.Delete;
 using DapperORM.Domain.Common.Result;
 using DapperORM.Domain.Constants;
 using DapperORM.Domain.Entities;
 using MediatR;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -30,7 +30,9 @@ namespace DapperORM.Application.Features.Commands.DeleteEvent
         public Task<IResult> Handle(DeleteProductCommandRequest request, CancellationToken cancellationToken)
         {
             Product product = mapper.Map<Product>(request);
-            EntityValidator.Validate(validator, product);
+            var result = validator.Validate(product);
+            if (result.Errors.Any())
+                return Task.FromResult<IResult>(new ErrorResult(result.Errors.First().ErrorMessage));
             productRepository.Delete(product);
             return Task.FromResult<IResult>(new SuccessResult(ResultMessages.Product_Deleted));
         }
